+++ tests/deprecated_api/test_remove_1-7.py
@@ -22,7 +22,6 @@ import torch
 
 from pytorch_lightning import Callback, Trainer
 from pytorch_lightning.callbacks.lr_monitor import LearningRateMonitor
-from pytorch_lightning.loggers import LoggerCollection
 from pytorch_lightning.overrides.distributed import IndexBatchSamplerWrapper
 from pytorch_lightning.plugins.environments import (
     KubeflowEnvironment,
@@ -83,11 +82,6 @@ def test_v1_7_0_deprecate_add_get_queue(tmpdir):
 def test_v1_7_0_lightning_logger_base_close(tmpdir):
     logger = CustomLogger()
     with pytest.deprecated_call(match="`Logger.close` method is deprecated in v1.5 and will be removed in v1.7."):
-        logger.close()
-    with pytest.deprecated_call(
-        match="`LoggerCollection.close` method is deprecated in v1.5 and will be removed in v1.7."
-    ):
-        logger = LoggerCollection([logger])
         logger.close()